add ByteBufferUtil
