+++ java/benchmark/src/main/java/org/apache/fury/benchmark/UserTypeDeserializeSuite.java
@@ -30,7 +30,7 @@ import org.apache.fury.benchmark.state.KryoState;
 import org.apache.fury.benchmark.state.ObjectType;
 import org.apache.fury.benchmark.state.ProtoBuffersState;
 import org.apache.fury.benchmark.state.ProtostuffState;
-import org.apache.fury.memory.Platform;
+import org.apache.fury.memory.ByteBufferUtil;
 import org.openjdk.jmh.Main;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
@@ -125,7 +125,7 @@ public class UserTypeDeserializeSuite {
 
   @Benchmark
   public Object flatbuffers_deserialize(FlatBuffersState.FlatBuffersUserTypeState state) {
-    Platform.clearBuffer(state.deserializedData);
+    ByteBufferUtil.clearBuffer(state.deserializedData);
     if (state.objectType == ObjectType.SAMPLE) {
       return FlatBuffersState.deserializeSample(state.deserializedData);
     } else {
